# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel TechTest API",
        "description": "Laravel TechTest API description",
        "contact": {
            "email": "keoz.mm@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0",
        "x-logo": {
            "url": "https://via.placeholder.com/190x90.png?text=L5-Swagger"
        }
    },
    "paths": {
...
        "/api/user/all": {
            "get": {
                "tags": [
                    "user"
                ],
                "summary": "Display a listing of the resource.",
                "operationId": "user@index",
                "responses": {
                    "200": {
                        "description": "Get resource",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request (URL or Params)"
                    },
                    "404": {
                        "description": "Resources not found"
                    }
                }
            }
        },
        "/api/user/create": {
            "post": {
                "tags": [
                    "user"
                ],
                "summary": "Store a newly created resource in storage.",
                "operationId": "user@store",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "firstname": {
                                        "description": "User firstname",
                                        "type": "string",
                                        "example": "Kevin Marroquin"
                                    },
                                    "lastname": {
                                        "description": "User lastname",
                                        "type": "string",
                                        "example": "Kevin Marroquin"
                                    },
                                    "email": {
                                        "description": "Email",
                                        "type": "string",
                                        "example": "kevin.marroquin@gmail.com"
                                    },
                                    "phone": {
                                        "description": "Phone",
                                        "type": "string",
                                        "example": "(503) 7958-8798"
                                    },
                                    "birthdate": {
                                        "description": "Birthdate",
                                        "type": "string",
                                        "example": "1998-03-19"
                                    },
                                    "gender": {
                                        "description": "Gender",
                                        "type": "string",
                                        "example": "Masculino"
                                    },
                                    "city": {
                                        "description": "City",
                                        "type": "string",
                                        "example": "San Vicente"
                                    },
                                    "country": {
                                        "description": "Country",
                                        "type": "string",
                                        "example": "El Salvador"
                                    },
                                    "address": {
                                        "description": "Address",
                                        "type": "string",
                                        "example": "Lorem ipsum dolor sit amet consectetur adipisicing elit."
                                    },
                                    "status": {
                                        "description": "Status",
                                        "type": "string",
                                        "example": "Activo"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Create resource",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request (URL or Params)"
                    },
                    "404": {
                        "description": "Resources not found"
                    },
                    "405": {
                        "description": "Validation exception"
                    }
                }
            }
        },
        "/api/user/filter/id/{id}": {
            "get": {
                "tags": [
                    "user"
                ],
                "summary": "Display the specified resource.",
                "operationId": "user@showById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "User to get by id",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Get resource",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request (URL or Params)"
                    },
                    "404": {
                        "description": "Resources not found"
                    }
                }
            }
        },
        "/api/user/filter/firstname/{firstname}": {
            "get": {
                "tags": [
                    "user"
                ],
                "summary": "Display the specified resource LIKE Firstname.",
                "operationId": "user@showByFirstname",
                "parameters": [
                    {
                        "name": "firstname",
                        "in": "path",
                        "description": "User to get by firstname",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Get resource",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request (URL or Params)"
                    },
                    "404": {
                        "description": "Resources not found"
                    }
                }
            }
        },
        "/api/user/filter/lastname/{lastname}": {
            "get": {
                "tags": [
                    "user"
                ],
                "summary": "Display the specified resource LIKE Lastname.",
                "operationId": "user@showByLastname",
                "parameters": [
                    {
                        "name": "lastname",
                        "in": "path",
                        "description": "User to get by lastname",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Get resource",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request (URL or Params)"
                    },
                    "404": {
                        "description": "Resources not found"
                    }
                }
            }
        },
        "/api/user/filter/email/{email}": {
            "get": {
                "tags": [
                    "user"
                ],
                "summary": "Display the specified resource LIKE Email.",
                "operationId": "user@showByEmail",
                "parameters": [
                    {
                        "name": "email",
                        "in": "path",
                        "description": "User to get by email",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Get resource",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request (URL or Params)"
                    },
                    "404": {
                        "description": "Resources not found"
                    }
                }
            }
        },
        "/api/user/filter/phone/{phone}": {
            "get": {
                "tags": [
                    "user"
                ],
                "summary": "Display the specified resource LIKE Phone.",
                "operationId": "user@showByphone",
                "parameters": [
                    {
                        "name": "phone",
                        "in": "path",
                        "description": "User to get by phone",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Get resource",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request (URL or Params)"
                    },
                    "404": {
                        "description": "Resources not found"
                    }
                }
            }
        },
...
    }
}